+++ features/src/codegen/features/domain/OneToOneAFooCodegen.java
@@ -14,9 +14,9 @@ import joist.domain.validation.rules.NotNull;
 public abstract class OneToOneAFooCodegen extends AbstractDomainObject {
 
     public static final OneToOneAFooQueries queries;
-    private Integer id = null;
+    private Long id = null;
     private String name = null;
-    private Integer version = null;
+    private Long version = null;
     private ForeignKeyListHolder<OneToOneAFoo, OneToOneABar> oneToOneABars = new ForeignKeyListHolder<OneToOneAFoo, OneToOneABar>((OneToOneAFoo) this, Aliases.oneToOneABar(), Aliases.oneToOneABar().oneToOneAFoo);
     protected Changed changed;
 
@@ -34,11 +34,11 @@ public abstract class OneToOneAFooCodegen extends AbstractDomainObject {
         this.addRule(new MaxLength<OneToOneAFoo>(Shims.name, 100));
     }
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.getChanged().record("id", this.id, id);
         this.id = id;
         if (UoW.isOpen()) {
@@ -55,7 +55,7 @@ public abstract class OneToOneAFooCodegen extends AbstractDomainObject {
         this.name = name;
     }
 
-    public Integer getVersion() {
+    public Long getVersion() {
         return this.version;
     }
 
@@ -99,11 +99,11 @@ public abstract class OneToOneAFooCodegen extends AbstractDomainObject {
     }
 
     static class Shims {
-        protected static final Shim<OneToOneAFoo, Integer> id = new Shim<OneToOneAFoo, Integer>() {
-            public void set(OneToOneAFoo instance, Integer id) {
+        protected static final Shim<OneToOneAFoo, Long> id = new Shim<OneToOneAFoo, Long>() {
+            public void set(OneToOneAFoo instance, Long id) {
                 ((OneToOneAFooCodegen) instance).id = id;
             }
-            public Integer get(OneToOneAFoo instance) {
+            public Long get(OneToOneAFoo instance) {
                 return ((OneToOneAFooCodegen) instance).id;
             }
             public String getName() {
@@ -121,11 +121,11 @@ public abstract class OneToOneAFooCodegen extends AbstractDomainObject {
                 return "name";
             }
         };
-        protected static final Shim<OneToOneAFoo, Integer> version = new Shim<OneToOneAFoo, Integer>() {
-            public void set(OneToOneAFoo instance, Integer version) {
+        protected static final Shim<OneToOneAFoo, Long> version = new Shim<OneToOneAFoo, Long>() {
+            public void set(OneToOneAFoo instance, Long version) {
                 ((OneToOneAFooCodegen) instance).version = version;
             }
-            public Integer get(OneToOneAFoo instance) {
+            public Long get(OneToOneAFoo instance) {
                 return ((OneToOneAFooCodegen) instance).version;
             }
             public String getName() {
@@ -141,8 +141,8 @@ public abstract class OneToOneAFooCodegen extends AbstractDomainObject {
         public boolean hasId() {
             return this.contains("id");
         }
-        public Integer getOriginalId() {
-            return (java.lang.Integer) this.getOriginal("id");
+        public Long getOriginalId() {
+            return (Long) this.getOriginal("id");
         }
         public boolean hasName() {
             return this.contains("name");
@@ -153,8 +153,8 @@ public abstract class OneToOneAFooCodegen extends AbstractDomainObject {
         public boolean hasVersion() {
             return this.contains("version");
         }
-        public Integer getOriginalVersion() {
-            return (java.lang.Integer) this.getOriginal("version");
+        public Long getOriginalVersion() {
+            return (Long) this.getOriginal("version");
         }
         public boolean hasOneToOneABars() {
             return this.contains("oneToOneABars");
